Added histogram mutex implementations
Given that we should first make sure histogram is behaving correctly (as this is very simple), these are two different implementations for the mutex strategy.

- 'histo_mutex' only uses one lock which controls the access to the entire array of bins at the end of the computation.
- 'histo_mutex_many' uses 256 locks, one for each bin.

NOTE: Both of these implementations are based on the assumption that we are not supposed to implement a naive approach for testing purposes.
In this case we should probably synchronize the access every time we read a pixel in the image.
However, this would simply deteriorate performance for no particular reason.