Pipesort working implementation
This is a complete refactoring of the pipesort algorithm according to the assignment specification.

We are now using bounded buffers (one per thread) for passing values and we have an additional worker called 'output' which only prints the values in order.

NOTE: The buffer size has been tested but not thouroghly. It seems like 2048 is a good value.