Reverted heat dissipation omp to our implementation
The omp based on reference is not working.
This version is an improvement over what we had for assignment 2 and it seems to perform better.

We don't have enough time to become acquainted with the ref implementation (I have tried and I can't get the threaded version to work without bugs).
It's better to stick with this one for comparisons as the pthread will also be based on this.